+++ src/unused-exports/settings.ts
@@ -1,13 +1,13 @@
-import * as vscode from 'vscode';
+import { workspace } from 'vscode';
 
 export function areMainExportsUsed(): boolean {
-  return vscode.workspace
+  return workspace
     .getConfiguration()
     .get('findUnusedExports.considerMainExportsUsed', false);
 }
 
 export function isResultExpanded(): boolean {
-  return vscode.workspace
+  return workspace
     .getConfiguration()
     .get('findUnusedExports.defaultResultExpanded', false);
 }